fix(ui): restore the frosted glass and extend the shell into the app
The glassmorphism was shipping broken and only in production. `.glass`
and `.scrim` each declared `backdrop-filter` before its `-webkit-`
prefixed twin, the minifier treated the pair as one property declared
twice and kept the last, and the standard property never reached the
bundle. Chromium computed `backdrop-filter: none`, so the chrome read as
flat translucency and the sidebar never blurred what sat behind it. Dev
looked correct throughout because dev does not minify, which is why this
survived a round of verification. Prefix first, standard last.

Texture is the other half of the same problem. A blur has nothing to
operate on over a flat fill, so the dot grid now backs the signed-in
shell as well as the public one, and the frost has something to smear.

The pointer trail and the footer follow it in. The safeguards and the
residency claims do not stop being true after login, and a screen that
simply ends is the emptiness this redesign is answering. Both shells now
share one reveal mechanism and one footer, so the classes lose their
`marketing-` prefix.

Also a thin, theme-aware scrollbar in both engines, and slightly wider
content columns.
